+++ src/main/java/com/authlete/jaxrs/spi/AuthorizationRequestHandlerSpiAdapter.java
@@ -78,4 +78,11 @@ public class AuthorizationRequestHandlerSpiAdapter implements AuthorizationReque
     {
         return null;
     }
+
+
+    @Override
+    public String[] getScopes()
+    {
+        return null;
+    }
 }